Cleaned models/ dir.
