+++ models/__init__.py
@@ -1,7 +1,7 @@
 from .arcnn import ARCNN
 from .mobilenet2 import MobileNetQA, MobileNetIR
 from .rrdbnet import RRDBNet
-from .prism import PrismLumaS4
-from .flare import FlareNet, FlareLuma, FlareChroma
+from .prism import Prism
+from .flare import Flare, FlareLuma, FlareChroma
 from .perceptual import Q1Net
-from .loss import MixedQ1MSELoss, Q1PerceptualLoss
+from .loss import MixedQ1MSELoss, Q1PerceptualLoss, CharbonnierLoss
