+++ models/dncnn.py
@@ -6,7 +6,6 @@ From paper:
 
 """
 
-import torch
 import torch.nn as nn
