+++ models/rrdbnet.py
@@ -1,7 +1,7 @@
 import torch
 import torch.nn as nn
 
-from .blocks import RRDB
+from .blocks import RRDB4
 
 
 class LumiNet(nn.Module):
@@ -18,7 +18,7 @@ class LumiNet(nn.Module):
         )
 
         self.body = nn.Sequential(
-            *[RRDB(64) for _ in range(num_blocks)]
+            *[RRDB4(64) for _ in range(num_blocks)]
         )
 
         self.decoder = nn.Sequential(
@@ -52,7 +52,7 @@ class ChromaNet(nn.Module):
         )
 
         self.body = nn.Sequential(
-            *[RRDB(64) for _ in range(num_blocks)]
+            *[RRDB4(64) for _ in range(num_blocks)]
         )
 
         self.decoder = nn.Sequential(
@@ -97,32 +97,3 @@ class RRDBNet(nn.Module):
         # Concatenate restored components
         result = torch.cat([Y_r, CbCr_r], dim=1)
         return result
-
-
-
-# class RRDBNet(nn.Module):
-#     def __init__(self, in_nc, out_nc, nf, nb, gc=32):
-#         super(RRDBNet, self).__init__()
-#         RRDB_block_f = functools.partial(RRDB, nf=nf, gc=gc)
-
-#         self.conv_first = nn.Conv2d(in_nc, nf, 3, 1, 1, bias=True)
-#         self.RRDB_trunk = make_layer(RRDB_block_f, nb)
-#         self.trunk_conv = nn.Conv2d(nf, nf, 3, 1, 1, bias=True)
-#         #### upsampling
-#         self.upconv1 = nn.Conv2d(nf, nf, 3, 1, 1, bias=True)
-#         self.upconv2 = nn.Conv2d(nf, nf, 3, 1, 1, bias=True)
-#         self.HRconv = nn.Conv2d(nf, nf, 3, 1, 1, bias=True)
-#         self.conv_last = nn.Conv2d(nf, out_nc, 3, 1, 1, bias=True)
-
-#         self.lrelu = nn.LeakyReLU(negative_slope=0.2, inplace=True)
-
-#     def forward(self, x):
-#         fea = self.conv_first(x)
-#         trunk = self.trunk_conv(self.RRDB_trunk(fea))
-#         fea = fea + trunk
-
-#         fea = self.lrelu(self.upconv1(F.interpolate(fea, scale_factor=2, mode='nearest')))
-#         fea = self.lrelu(self.upconv2(F.interpolate(fea, scale_factor=2, mode='nearest')))
-#         out = self.conv_last(self.lrelu(self.HRconv(fea)))
-
-#         return out
